packages/listener/pyproject.toml
[project]
name = "listener"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "massive>=1.0.0",
    "python-dotenv>=1.0.0",
    "domain",
    "store",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/listener"]
